Add linklist to unitview
